backend/requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.19.0
Pillow>=10.0.0
websockets>=13.0
httpx>=0.27.0
pydantic>=2.0.0
